+++ CHANGELOG.md
@@ -131,3 +131,4 @@ was removed or the new one was added to folder
 - Scene Man is now based on smart pointers
 - File explorer now allows for easy reset of selected path
 - Few bug fixes in path maker
+- Text Render Type and Render Type now uses Enums